Make `BuildError` under `error::operation` in `smithy-types`
This commit addresses https://github.com/awslabs/smithy-rs/pull/3032#discussion_r1349138255
