+++ rust-runtime/aws-smithy-http/src/operation.rs
@@ -16,7 +16,9 @@ use std::ops::{Deref, DerefMut};
 //TODO(runtimeCratesVersioningCleanup): Re-point those who use the following reexport to
 // directly depend on `aws_smithy_types` and remove the reexport below.
 /// Errors for operations
-pub use aws_smithy_types::operation::error;
+pub mod error {
+    pub use aws_smithy_types::error::operation::{BuildError, SerializationError};
+}
 
 /// Metadata attached to an [`Operation`] that identifies the API being called.
 #[derive(Clone, Debug)]
